fix: add JsonCreator to fix serialization error
