--- src/main/java/nextstep/auth/TokenRequest.java
@@ -1,19 +1,0 @@
-package nextstep.auth;
-
-public class TokenRequest {
-    private String username;
-    private String password;
-
-    public TokenRequest(String username, String password) {
-        this.username = username;
-        this.password = password;
-    }
-
-    public String getUsername() {
-        return username;
-    }
-
-    public String getPassword() {
-        return password;
-    }
-}
